Migrate method long to int
* Migrate the annotation id from a long to int.
* Updated all related tests and related files.

+++ src/main/java/org/jgroups/raft/StateMachineWrite.java
@@ -36,7 +36,7 @@ public @interface StateMachineWrite {
      *
      * @return the operation identifier
      */
-    long id();
+    int id();
 
     /**
      * Version number for backwards compatibility.
