+++ src/main/java/org/jgroups/raft/internal/registry/CommandMetadata.java
@@ -5,7 +5,7 @@ import org.jgroups.raft.internal.command.JRaftCommand;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 
-record CommandMetadata(Object stateMachine, long id, int version, Method method, CommandSchema inputSchema, CommandSchema outputSchema) implements ReplicatedMethodWrapper {
+record CommandMetadata(Object stateMachine, int id, int version, Method method, CommandSchema inputSchema, CommandSchema outputSchema) implements ReplicatedMethodWrapper {
 
     void validate(Method method, JRaftCommand command) {
         if (id != command.id()) {
